+++ SigkSens/SigkSens.ino
@@ -127,7 +127,6 @@ Reactduino app([] () {
   setupWifi();
   loadConfig();
   setupDiscovery();
-  setupHTTP();
   setupWebSocket();
   setupSignalK();
 
@@ -155,6 +154,10 @@ Reactduino app([] () {
     saveConfig();
   }
   
+  // call http last so that we can call any needed callbacks.
+  setupHTTP();
+
+
   Serial.printf("Ready!\n");
 
   app.repeat(SLOW_LOOP_DELAY, &slow_loop);
