+++ SigkSens/src/net/httpd.cpp
@@ -20,15 +20,6 @@ extern "C" {
 #include "../sensors/sensorStorage.h"
 #include "../services/configReset.h"
 
-#ifdef ENABLE_MPU
-  #include "../sensors/mpu9250/mpu9250.h"
-#endif
-
-
-
-
-
-
 // SSDP related stuff
 
 //SSDP properties
@@ -215,24 +206,6 @@ void httpSetSignalKToken(AsyncWebServerRequest *request) {
 }
 
 
-#ifdef ENABLE_MPU
-void httpMpuCalAccelGyro(AsyncWebServerRequest *request) {
-  runAccelGyroCal();
-  request->send(200, "application/json", "{ \"success\": true }");
-}
-
-void httpMpuCalMagStart(AsyncWebServerRequest *request) {
-  runMagCalStart();
-  request->send(200, "application/json", "{ \"success\": true }");
-}
-
-void httpMpuCalMagStop(AsyncWebServerRequest *request) {
-  runMagCalStop();
-  request->send(200, "application/json", "{ \"success\": true }");
-}
-
-#endif
-
 void httpGetSensorInfo(AsyncWebServerRequest *request) {
   AsyncJsonResponse * response = new AsyncJsonResponse();
   JsonObject& json = response->getRoot();
@@ -395,14 +368,6 @@ void setupHTTP() {
   server.on("/setTimerDelay", HTTP_GET, httpSetTimerDelay);
   server.on("/setNewHostname", HTTP_GET, httpNewHostname);
 
-
-  #ifdef ENABLE_MPU
-  server.on("/mpuCalAccelGyro", HTTP_GET, httpMpuCalAccelGyro);
-  server.on("/mpuCalMagStart", HTTP_GET, httpMpuCalMagStart);
-  server.on("/mpuCalMagStop", HTTP_GET, httpMpuCalMagStop);
-  #endif
-
-
   server.on("/setSignalKHost", HTTP_GET, httpSetSignalKHost);
   server.on("/setSignalKPort", HTTP_GET, httpSetSignalKPort);
   server.on("/setSignalKPath", HTTP_GET, httpSetSignalKPath);
@@ -433,6 +398,12 @@ void setupHTTP() {
   server.on("/signalk/", HTTP_GET, httpSignalKEndpoints);
   
   server.on("/resetConfig", HTTP_GET, httpResetConfig);
+
+  //setup sensor callbacks
+  sensorStorageForEach([&](SensorInfo* si) {
+    Serial.print("Calling callback for"); Serial.println(si->address);
+    si->setupWebServerHooks(server);
+  });
 
   server.begin();
 }
