+++ SigkSens/src/sensors/sensofInfo.cpp
@@ -4,5 +4,7 @@
 
 #include "sensorInfo.h"
 
+void SensorInfo::setupWebServerHooks(AsyncWebServer& server){ }
+
 
 fromJsonFunc fromJson[(int)SensorType::SensorType_MAX+1];
